fix a couple of verbs
These are minor modifications. 

The codepen embedded at the top might be better described through a link tag, as in `We will be creating [this](http://codepen.io/wesbos/pen/KgpNjJ/)`, but it's more of a design choice. The embed will certainly be better when the markdown file is used in the actual blog.